move profiles receiver into the main receiver package
